separate into files and build with browserify
