+++ API.md
@@ -2,7 +2,6 @@
 
 ### Table of Contents
 
--   [FontCanvas](#fontcanvas)
 -   [VectorTiles](#vectortiles)
 -   [initialize](#initialize)
 -   [search](#search)
@@ -11,25 +10,9 @@
 -   [restyleByProperty](#restylebyproperty)
 -   [setFeatureStyle](#setfeaturestyle)
 -   [resetFeatureStyle](#resetfeaturestyle)
--   [getFeatureGroup](#getfeaturegroup)
 -   [getLayer](#getlayer)
 -   [getGeoJSON](#getgeojson)
 -   [removeFeature](#removefeature)
-
-## FontCanvas
-
-**Extends Canvas**
-
-A canvas renderer that can draw fonts.
-Useful for icon fonts.
-
-**Examples**
-
-```javascript
-var map = L.map('map', {
-  renderer: new L.FontCanvas()
-});
-```
 
 ## VectorTiles
 
@@ -123,13 +106,6 @@ Revert a feature to its origin style.
 **Parameters**
 
 -   `id` **[string](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String)** 
-
-## getFeatureGroup
-
-Returns the feature group that holds all features in the GridLayer
-intended for use with Leaflet.Draw
-
-Returns **L.FeatureGroup** 
 
 ## getLayer
