+++ package.json
@@ -7,8 +7,10 @@
     "example": "example"
   },
   "scripts": {
+    "build": "./node_modules/.bin/browserify Leaflet.VectorTiles.js -o dist/leaflet.vector-tiles.js -t [ babelify --presets [ es2015 ] ]",
+    "build-dev": "./node_modules/.bin/browserify leaflet.vector-tiles.js -o dist/leaflet.vector-tiles.js -t [ babelify --presets [ es2015 ] ] --debug",
     "build-docs": "./node_modules/.bin/documentation build Leaflet.VectorTiles.js -f md > API.md",
-    "lint": "./node_modules/.bin/eslint Leaflet.VectorTiles.js",
+    "lint": "./node_modules/.bin/eslint Leaflet.VectorTiles.js tile.js feature.js Leaflet.FontCanvas.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -22,9 +24,16 @@
   },
   "homepage": "https://github.com/NBTSolutions/Leaflet.VectorTiles#readme",
   "devDependencies": {
+    "babel-preset-es2015": "^6.24.1",
+    "babelify": "^7.3.0",
+    "browserify": "^14.3.0",
     "documentation": "^4.0.0-rc.1",
     "eslint": "^3.19.0",
     "eslint-config-airbnb-base": "^11.2.0",
     "eslint-plugin-import": "^2.2.0"
+  },
+  "dependencies": {
+    "@turf/bbox": "^4.3.0",
+    "rbush": "^2.0.1"
   }
 }
